Add check for no DWS Storages
Problem: When there are no `Storages` available in k8s, no list of rabbits
is available. This essentially results in a KeyError since the rabbit
dicts are empty:

```
2024-12-16 10:06:18,840 - ERROR - Error in responding to dws.setup RPC for 14489491083362304: 'rabbit-compute-2'
```

Add a check to make sure there are `Storages` available, if not, then
exit since that means there are no rabbits available, and no workflows
will be able to run.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>